[DF-11625] warnings fixed
Former-commit-id: 1f0d4e26b86432ad42bf43f72e70d6d6cb219cab

+++ Sources/External/cef/Sources/CEFWebPageRender.cpp
@@ -158,7 +158,7 @@ void CEFWebPageRender::ResetCursor()
 
 bool CEFWebPageRender::GetViewRect(CefRefPtr<CefBrowser> browser, CefRect& rect)
 {
-    rect = CefRect(0, 0, logicalViewSize.dx, logicalViewSize.dy);
+    rect = CefRect(0, 0, static_cast<int>(logicalViewSize.dx), static_cast<int>(logicalViewSize.dy));
     return true;
 }
